+++ src/runtime/server/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../.nuxt/tsconfig.server.json",
+}
